added build script for precomputed things
