+++ Cargo.toml
@@ -5,6 +5,9 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[build-dependencies]
+rand = "0.8.5"
+
 [dependencies]
 rand = "0.8.5"
 serde = { version = "1.0.164", features = ["derive"] }
