+++ src/precomputed/precomputed.rs
@@ -1,0 +1,261 @@
+pub static KNIGHT_MOVES: [u64; 64] = [
+0x0000000000020400u64, 
+0x0000000000050800u64, 
+0x00000000000A1100u64, 
+0x0000000000142200u64, 
+0x0000000000284400u64, 
+0x0000000000508800u64, 
+0x0000000000A01000u64, 
+0x0000000000402000u64, 
+0x0000000002040004u64, 
+0x0000000005080008u64, 
+0x000000000A110011u64, 
+0x0000000014220022u64, 
+0x0000000028440044u64, 
+0x0000000050880088u64, 
+0x00000000A0100010u64, 
+0x0000000040200020u64, 
+0x0000000204000402u64, 
+0x0000000508000805u64, 
+0x0000000A1100110Au64, 
+0x0000001422002214u64, 
+0x0000002844004428u64, 
+0x0000005088008850u64, 
+0x000000A0100010A0u64, 
+0x0000004020002040u64, 
+0x0000020400040200u64, 
+0x0000050800080500u64, 
+0x00000A1100110A00u64, 
+0x0000142200221400u64, 
+0x0000284400442800u64, 
+0x0000508800885000u64, 
+0x0000A0100010A000u64, 
+0x0000402000204000u64, 
+0x0002040004020000u64, 
+0x0005080008050000u64, 
+0x000A1100110A0000u64, 
+0x0014220022140000u64, 
+0x0028440044280000u64, 
+0x0050880088500000u64, 
+0x00A0100010A00000u64, 
+0x0040200020400000u64, 
+0x0204000402000000u64, 
+0x0508000805000000u64, 
+0x0A1100110A000000u64, 
+0x1422002214000000u64, 
+0x2844004428000000u64, 
+0x5088008850000000u64, 
+0xA0100010A0000000u64, 
+0x4020002040000000u64, 
+0x0400040200000000u64, 
+0x0800080500000000u64, 
+0x1100110A00000000u64, 
+0x2200221400000000u64, 
+0x4400442800000000u64, 
+0x8800885000000000u64, 
+0x100010A000000000u64, 
+0x2000204000000000u64, 
+0x0004020000000000u64, 
+0x0008050000000000u64, 
+0x00110A0000000000u64, 
+0x0022140000000000u64, 
+0x0044280000000000u64, 
+0x0088500000000000u64, 
+0x0010A00000000000u64, 
+0x0020400000000000u64];
+pub static KING_MOVES: [u64; 64] = [
+0x0000000000000302u64, 
+0x0000000000000705u64, 
+0x0000000000000E0Au64, 
+0x0000000000001C14u64, 
+0x0000000000003828u64, 
+0x0000000000007050u64, 
+0x000000000000E0A0u64, 
+0x000000000000C040u64, 
+0x0000000000030203u64, 
+0x0000000000070507u64, 
+0x00000000000E0A0Eu64, 
+0x00000000001C141Cu64, 
+0x0000000000382838u64, 
+0x0000000000705070u64, 
+0x0000000000E0A0E0u64, 
+0x0000000000C040C0u64, 
+0x0000000003020300u64, 
+0x0000000007050700u64, 
+0x000000000E0A0E00u64, 
+0x000000001C141C00u64, 
+0x0000000038283800u64, 
+0x0000000070507000u64, 
+0x00000000E0A0E000u64, 
+0x00000000C040C000u64, 
+0x0000000302030000u64, 
+0x0000000705070000u64, 
+0x0000000E0A0E0000u64, 
+0x0000001C141C0000u64, 
+0x0000003828380000u64, 
+0x0000007050700000u64, 
+0x000000E0A0E00000u64, 
+0x000000C040C00000u64, 
+0x0000030203000000u64, 
+0x0000070507000000u64, 
+0x00000E0A0E000000u64, 
+0x00001C141C000000u64, 
+0x0000382838000000u64, 
+0x0000705070000000u64, 
+0x0000E0A0E0000000u64, 
+0x0000C040C0000000u64, 
+0x0003020300000000u64, 
+0x0007050700000000u64, 
+0x000E0A0E00000000u64, 
+0x001C141C00000000u64, 
+0x0038283800000000u64, 
+0x0070507000000000u64, 
+0x00E0A0E000000000u64, 
+0x00C040C000000000u64, 
+0x0302030000000000u64, 
+0x0705070000000000u64, 
+0x0E0A0E0000000000u64, 
+0x1C141C0000000000u64, 
+0x3828380000000000u64, 
+0x7050700000000000u64, 
+0xE0A0E00000000000u64, 
+0xC040C00000000000u64, 
+0x0203000000000000u64, 
+0x0507000000000000u64, 
+0x0A0E000000000000u64, 
+0x141C000000000000u64, 
+0x2838000000000000u64, 
+0x5070000000000000u64, 
+0xA0E0000000000000u64, 
+0x40C0000000000000u64];
+pub static ROOK_MASK: [u64; 64] = [
+    72340172838076926,
+    144680345676153597,
+    289360691352306939,
+    578721382704613623,
+    1157442765409226991,
+    2314885530818453727,
+    4629771061636907199,
+    9259542123273814143,
+    72340172838141441,
+    144680345676217602,
+    289360691352369924,
+    578721382704674568,
+    1157442765409283856,
+    2314885530818502432,
+    4629771061636939584,
+    9259542123273813888,
+    72340172854657281,
+    144680345692602882,
+    289360691368494084,
+    578721382720276488,
+    1157442765423841296,
+    2314885530830970912,
+    4629771061645230144,
+    9259542123273748608,
+    72340177082712321,
+    144680349887234562,
+    289360695496279044,
+    578721386714368008,
+    1157442769150545936,
+    2314885534022901792,
+    4629771063767613504,
+    9259542123257036928,
+    72341259464802561,
+    144681423712944642,
+    289361752209228804,
+    578722409201797128,
+    1157443723186933776,
+    2314886351157207072,
+    4629771607097753664,
+    9259542118978846848,
+    72618349279904001,
+    144956323094725122,
+    289632270724367364,
+    578984165983651848,
+    1157687956502220816,
+    2315095537539358752,
+    4629910699613634624,
+    9259541023762186368,
+    143553341945872641,
+    215330564830528002,
+    358885010599838724,
+    645993902138460168,
+    1220211685215703056,
+    2368647251370188832,
+    4665518383679160384,
+    9259260648297103488,
+    18302911464433844481,
+    18231136449196065282,
+    18087586418720506884,
+    17800486357769390088,
+    17226286235867156496,
+    16077885992062689312,
+    13781085504453754944,
+    9187484529235886208,
+];pub static BISHOP_MASK: [u64; 64] = [
+    9241421688590303744,
+    36099303471056128,
+    141012904249856,
+    550848566272,
+    6480472064,
+    1108177604608,
+    283691315142656,
+    72624976668147712,
+    4620710844295151618,
+    9241421688590368773,
+    36099303487963146,
+    141017232965652,
+    1659000848424,
+    283693466779728,
+    72624976676520096,
+    145249953336262720,
+    2310355422147510788,
+    4620710844311799048,
+    9241421692918565393,
+    36100411639206946,
+    424704217196612,
+    72625527495610504,
+    145249955479592976,
+    290499906664153120,
+    1155177711057110024,
+    2310355426409252880,
+    4620711952330133792,
+    9241705379636978241,
+    108724279602332802,
+    145390965166737412,
+    290500455356698632,
+    580999811184992272,
+    577588851267340304,
+    1155178802063085600,
+    2310639079102947392,
+    4693335752243822976,
+    9386671504487645697,
+    326598935265674242,
+    581140276476643332,
+    1161999073681608712,
+    288793334762704928,
+    577868148797087808,
+    1227793891648880768,
+    2455587783297826816,
+    4911175566595588352,
+    9822351133174399489,
+    1197958188344280066,
+    2323857683139004420,
+    144117404414255168,
+    360293502378066048,
+    720587009051099136,
+    1441174018118909952,
+    2882348036221108224,
+    5764696068147249408,
+    11529391036782871041,
+    4611756524879479810,
+    567382630219904,
+    1416240237150208,
+    2833579985862656,
+    5667164249915392,
+    11334324221640704,
+    22667548931719168,
+    45053622886727936,
+    18049651735527937,
+];
